Use concat on split words for sentence
To ensure no multiple word strings get added to the sentence via
random_word, split strings as they come in and concat them to the
array. Need to look into the handling of random_word to see if theres a
way to get rid of the mostly wasteful split operation.